App Load - download all condition and slide data on app load
After my walk, I decided that it's probably easiest to look at
downloading all of the data when the app loads. This is done in chunks
by condition, so that there is still time for the app to respond to
interations. I don't think that it will really take too long though as
the data for each condition and each slide is minimal.