ErrorHandling - set up error handling provider
This owns the instance of Sentry that the rest of the app can access via
a hook. I don't know if this is going to be an easier way of managing
error handling, but I've wanted to set something up like this for ages
with this kind of dependency injection, so let's see how it goes!