test: Fix gnarly broken `App` test
Man this bug was insane - I don't fully understand what was happening,
but it was caused by my not returning an array from a mocked function
that should have been returning an array. Instead it was returning
`undefined` which should have been erroring in the reducer, but for some
reason it wasn't erroring until it had run it four times, at which point
it would just unmount the component.

I don't know whether it was `jest` or `react` or `react-test-renderer`
catching the error, but someone was keeping it to themselves, so I was
totally chasing my tail on this one!