feat(App): make "browse" the landing page of the app
Navigation had to be rethought to work from here. Also, the tests had to
be updated to take into account the new layout of the app. Not great
that all of the tests depend on the flow of the app. They're more like
e2e tests. But actually they didn't take too long to update, so could be
worse!